In SingleMultiply, cause the weight to be a real float, not the result of an integer division cast out to a float.

+++ streams/apps/benchmarks/fir/streamit/FIRfine.java
@@ -63,7 +63,7 @@ class SingleMultiply extends Filter
 
     float W;
     public void init(final int i) {
-	W = 2*i*i/(i+1);
+	W = 2*i*i/((float)i+1);
 	this.input = new Channel(Float.TYPE, 2, 2);
 	this.output = new Channel(Float.TYPE, 2);
     }
